pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tatha"
version = "0.1.0"
description = "Tatha 主仓能力层：简历解析、匹配、诗人/诗词 RAG，基于九库统一模型与文档链路"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # 1. LiteLLM - 统一多平台模型调用
    "litellm>=1.0.0",
    # 2. MarkItDown - 多格式文档转 Markdown
    "markitdown>=0.0.1",
    # 3. LlamaIndex - 文档索引与 RAG
    "llama-index>=0.10.0",
    "llama-index-core>=0.10.0",
    "llama-index-embeddings-openai",
    "llama-index-vector-stores-faiss",
    # 4. PydanticAI - 类型安全智能体
    "pydantic-ai>=0.0.1",
    "pydantic>=2.0",
    # 5. Marvin - AI 封装为函数
    "marvin>=2.0.0",
    # 6. Haystack - 检索流水线（可选）
    "haystack-ai>=2.0.0",
    # 7. tiktoken - Token 计数
    "tiktoken>=0.5.0",
    # 8. FAISS - 向量检索（LlamaIndex 已可集成，单独装以便直接使用）
    "faiss-cpu>=1.7.0",
    # 9. Pydantic Evals - 提示词回归测试
    "pydantic-evals>=0.0.1",
    # API 与工具
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "python-dotenv>=1.0.0",
    "pydantic-ai-litellm>=0.2.3",
    "llama-index-embeddings-huggingface>=0.6.1",
    "llama-index-llms-litellm>=0.6.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
]

[tool.setuptools.packages.find]
where = ["src"]
